Initial Commit
This is a pilot work for a larger project on Adversarial Machine Learning. This work utilizes the OpenCV Haarcascade library to detect eye area of human face. It highlights the eye region and generates a mask of the email. The White portion of mask denotes the eye region while the rest of image is blacked out.